+++ panweave-zcl/src/lib.rs
@@ -30,6 +30,7 @@ pub mod clusters;
 pub mod frame;
 pub mod global;
 pub mod layer;
+pub mod requirements;
 pub mod structured;
 pub mod types;